% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SIRsolver}
\alias{SIRsolver}
\title{Run a single SIR simulation a la eqns. 6, 7 from Gandon
et al 2001 Nature to optimize virulence in the face of
different medical interventions}
\usage{
SIRsolver(
  density_init = 100,
  f = 0.3,
  lambda = 25,
  delta = 1,
  b1 = 1,
  b2 = 0.5,
  c1 = 1,
  c2 = 0.2,
  r1 = 0,
  r2 = 0,
  r3 = 0,
  r4 = 0,
  sigma = 0,
  maxt = 100000000L
)
}
\arguments{
\item{density_init}{initial numbers (equal for susceptible and infected) (floating-point)}

\item{f}{fraction vaccinated (floating-point)}

\item{lambda}{birth rate (floating-point)}

\item{delta}{parasite-independent mortality rate (floating-point)}

\item{b1}{intercept of trade-off curve virulence-induced mortality vs transmission (floating-point)}

\item{b2}{power of trade-off curve virulence-induced mortality vs transmission (floating-point)}

\item{c1}{intercept of trade-off curve virulence-induced mortality vs recovery (floating-point)}

\item{c2}{power of trade-off curve virulence-induced mortality vs recovery (floating-point)}

\item{r1}{anti-infection immunity (floating-point value)}

\item{r2}{anti-growth rate immunity (floating-point value)}

\item{r3}{transmission-blocking immunity (floating-point value)}

\item{r4}{anti-toxin immunity (floating-point value)}
}
\value{
A \code{list} that contains the values of the equilibrium 
    densities x, x', y, y' and virulence alpha
}
\description{
Run a single SIR simulation a la eqns. 6, 7 from Gandon
et al 2001 Nature to optimize virulence in the face of
different medical interventions
}
\examples{
    # test run of a simulation
    list.result <- SIRsolver(
                        density_init=100,
                        f=0.2,
                        lambda=25,
                        delta=1,
                        b1=1,
                        b2=0.5,
                        c1=1,
                        c2=0.2,
                        r1=0,
                        r2=0,
                        r3=0,
                        r4=0,
                        sigma=0,
                        maxt=50000)

}
